requirements.txt
python-dotenv
sounddevice
numpy
requests
keyboard
pyautogui
pynput 